pyproject.toml
# --- BUILD SYSTEM ---

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

# --- PROJECT ---

[project]
name = "browser-use-mcp"
version = "0.0.1"
description = "A Browser Use MCP server, built with the Dedalus MCP framework."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Dedalus Labs", email = "oss@dedaluslabs.ai" },
]
maintainers = [
    { name = "Dedalus Labs", email = "oss@dedaluslabs.ai" },
]
keywords = ["ai", "mcp", "agents", "dedalus", "browser-use", "cloud"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "dedalus-mcp>=0.7.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvloop>=0.22.1; platform_system != 'Windows'",
]

# --- PROJECT URLS ---

[project.urls]
Homepage = "https://github.com/dedalus-labs/browser-use-mcp"
Repository = "https://github.com/dedalus-labs/browser-use-mcp"
Issues = "https://github.com/dedalus-labs/browser-use-mcp/issues"

# --- DEPENDENCY GROUPS ---

[dependency-groups]
runtime = [
    "pydantic>=2.12.5",
    "uvloop>=0.22.1; platform_system != 'Windows'",
]

test = [
    "anyio>=4.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

lint = ["ruff>=0.13.3", "pre-commit>=4.3.0", "ty>=0.0.1a21"]

debug = ["pdbpp>=0.10.3", "icecream>=2.1.8"]

dev = [
    { include-group = "runtime" },
    { include-group = "test" },
    { include-group = "lint" },
    { include-group = "debug" },
]

# --- UV ---

[tool.uv]
preview = true
package = false
default-groups = ["runtime"]
cache-dir = ".cache/uv"
compile-bytecode = true
link-mode = "hardlink"
environments = ["sys_platform == 'linux'", "sys_platform == 'darwin'"]
concurrent-downloads = 50
concurrent-builds = 4
concurrent-installs = 4
resolution = "highest"
prerelease = "if-necessary"
fork-strategy = "fewest"
index-strategy = "first-index"
python-preference = "managed"
python-downloads = "automatic"

[tool.uv.sources]

[tool.uv.build-backend]
source-exclude = [
    "dist/**",
    "**/__pycache__/**",
    "**/build/**",
    "**/.pytest_cache/**",
    "**/.mypy_cache/**",
]

# --- RUFF ---

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["build/", "dist/", ".venv/", "__pycache__/"]

# --- RUFF FORMATTING ---

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72
skip-magic-trailing-comma = false
quote-style = "double"
indent-style = "space"

# --- RUFF LINTING ---

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "UP",
    "B",
    "C4",
    "SIM",
    "RUF",
    "D",
    "N",
    "ANN",
    "TCH",
    "S",
    "BLE",
    "TRY",
    "RSE",
    "RET",
    "PERF",
    "FLY",
    "FURB",
    "ARG",
    "COM",
    "DTZ",
    "EM",
    "ERA",
    "EXE",
    "FBT",
    "G",
    "ICN",
    "INP",
    "INT",
    "ISC",
    "LOG",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "SLF",
    "SLOT",
    "T20",
    "TID",
]

ignore = [
    "COM812",
    "E501",
    "ISC001",
    "E402",
    "TID252",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D107",
    "S101",
    "T201",
    "RET504",
    "FBT001",
    "FBT002",
    "PLR0913",
]

unfixable = ["B", "TRY", "RUF"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["browser", "tools"]
force-single-line = false
force-sort-within-sections = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
split-on-trailing-comma = true
lines-after-imports = 2
relative-imports-order = "closest-to-furthest"

# --- PYTEST ---

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]
testpaths = ["tests"]

# --- TY ---

[tool.ty.rules]
invalid-assignment = "error"
invalid-return-type = "error"
invalid-argument-type = "error"
missing-argument = "error"
too-many-positional-arguments = "error"
unresolved-reference = "error"
invalid-base = "error"
conflicting-metaclass = "error"
subclass-of-final-class = "error"
undefined-reveal = "warn"
division-by-zero = "ignore"
unused-ignore-comment = "ignore"

[tool.ty.environment]
root = ["src"]
python-platform = "darwin"
python-version = "3.11"

[tool.ty.terminal]
output-format = "concise"
error-on-warning = false

[tool.ty.src]
include = ["**/*.py", "**/*.pyi"]
exclude = [
    "build/**",
    "dist/**",
    ".venv/**",
    "__pycache__/**",
    ".pytest_cache/**",
    "src/_client.py",
    "tests/**",
]
respect-ignore-files = true
